requirements.txt
torch
transformers
pillow
pandas
pyyaml
matplotlib
plotly
datasets
requests
scikit-learn
